% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{adjust_image_path_windows}
\alias{adjust_image_path_windows}
\title{Add /mnt/c/ to path to make it compatible with WSL.}
\usage{
adjust_image_path_windows(image_path)
}
\arguments{
\item{image_path}{}
}
\value{
Adjusted path.
}
\description{
Add /mnt/c/ to path to make it compatible with WSL.
}
