% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{alline_image}
\alias{alline_image}
\title{Aligns image to reference image.}
\usage{
alline_image(
  base_path,
  in_file,
  out_file,
  reference_file,
  emask = NULL,
  extra_alline_opts = NULL,
  overwrite = TRUE
)

alline_image(
  base_path,
  in_file,
  out_file,
  reference_file,
  emask = NULL,
  extra_alline_opts = NULL,
  overwrite = TRUE
)

alline_image(
  base_path,
  in_file,
  out_file,
  reference_file,
  emask = NULL,
  extra_alline_opts = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{reference_file}{}
}
\value{





}
\description{
Aligns image to reference image.

Aligns image to reference image.

Aligns image to reference image.
}
