% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness_functions.R
\name{binary_clever_score}
\alias{binary_clever_score}
\title{Clever score on l2 norm for binary classification with sigmoid activation.}
\usage{
binary_clever_score(model_build_fn, model_params, path_to_weights, x_test)
}
\arguments{
\item{model_build_fn}{Function to build model.}

\item{model_params}{List of model parameters.}

\item{path_to_weights}{Absolute path to model weights.}

\item{x_test}{Examples on which to calculate score.}
}
\value{
Lit of clever score for each test example.
}
\description{
Clever score on l2 norm for binary classification with sigmoid activation.
}
