% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{build_afni_cmd}
\alias{build_afni_cmd}
\title{Builds command to run in AFNI.}
\usage{
build_afni_cmd(afni_command, afni_opts = NULL)
}
\arguments{
\item{afni_opts}{}
}
\value{

}
\description{
Builds command to run in AFNI.
}
