% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_DeepMedic.R
\name{build_DeepMedic}
\alias{build_DeepMedic}
\title{Function to build the DeepMedic Network from scratch so it can be customized}
\usage{
build_DeepMedic(model_params)
}
\arguments{
\item{model_params}{all 3 spatial dimensions of input shape must be even; must
also specify downsamp factor for downsampled pathway and kernel size of conv layers.}
}
\value{

}
\description{
Function to build the DeepMedic Network from scratch so it can be customized
}
