% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ResNet.R
\name{build_ResNet}
\alias{build_ResNet}
\title{Builds ResNet 50}
\usage{
build_ResNet(numResNet, model_params)
}
\arguments{
\item{numResNet}{Which resnet to use (ResNet50, ResNet101, ResNet152)}

\item{model_params}{List of model parameters.}
}
\value{
ResNet model
}
\description{
Builds ResNet 50
}
