% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_UResNet.R
\name{build_UResNet}
\alias{build_UResNet}
\title{Build Residual U-Net}
\usage{
build_UResNet(model_params)

build_UResNet(model_params)

build_UResNet(model_params)
}
\arguments{
\item{model_params}{Model params list, must add batch size when making list.
Also, x,y,z dims of input_shape must be even.}
}
\value{





}
\description{
Build Residual U-Net

Build Residual U-Net

Build Residual U-Net
}
