% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{check_image_max}
\alias{check_image_max}
\title{Get max value of nii.}
\usage{
check_image_max(image_path)

check_image_max(image_path)

check_image_max(image_path)
}
\arguments{
\item{image_path}{}
}
\value{
image max.

image max.

image max.
}
\description{
Get max value of nii.

Get max value of nii.

Get max value of nii.
}
