% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{chunk_image}
\alias{chunk_image}
\title{Function to chunk image for fitting into memory.}
\usage{
chunk_image(num_chunks, chunk_size)
}
\value{

}
\description{
Function to chunk image for fitting into memory.
}
