% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{dice_loss}
\alias{dice_loss}
\title{Function to define dice loss for networks}
\usage{
dice_loss(y_true, y_pred)
}
\arguments{
\item{y_pred}{}
}
\value{

}
\description{
Function to define dice loss for networks
}
