% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{extract_highest_diffusion_shell}
\alias{extract_highest_diffusion_shell}
\title{Title}
\usage{
extract_highest_diffusion_shell(in_file, out_file)
}
\arguments{
\item{out_file}{}
}
\value{

}
\description{
Title
}
