% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_functions.R
\name{gradcam_pp}
\alias{gradcam_pp}
\title{3dGradCAM++}
\usage{
gradcam_pp(test_sample, ref_img, write_imgs = NULL)
}
\arguments{
\item{test_sample}{Example to use for grad cam.}

\item{ref_img}{Path to image used as reference for preprocessing.}

\item{write_imgs}{If not NULL, where to write nifti for cam and orig imgs.}
}
\value{
Aligned gradcam and orig image as nii objects.
}
\description{
3dGradCAM++
}
