% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hcf}
\alias{hcf}
\title{Find highest common factor between 2 numbers, useful for chunking rectangular matrix.}
\usage{
hcf(x, y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Find highest common factor between 2 numbers, useful for chunking rectangular matrix.
}
