% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_functions.R
\name{jaccard_loss}
\alias{jaccard_loss}
\title{Jaccard loss, better for unbalanced datasets.}
\usage{
jaccard_loss(y_true, y_pred)
}
\arguments{
\item{smooth}{}
}
\value{

}
\description{
Jaccard loss, better for unbalanced datasets.
}
