% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_functions.R
\name{model_modifier}
\alias{model_modifier}
\title{Builds model modifier for GradCam++}
\usage{
model_modifier(m)
}
\arguments{
\item{m}{}
}
\value{
model_modifier
}
\description{
Builds model modifier for GradCam++
}
