% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{resample_image_afni}
\alias{resample_image_afni}
\title{Resample image to a particular grid resolution.}
\usage{
resample_image_afni(in_file, out_file, new_dims, extra_opts = NULL)
}
\arguments{
\item{in_file}{Absolute path to file to be reasampled}

\item{out_file}{Absolute path to save file destination.}

\item{new_dims}{Vector of new dimensions c(x,y,z)}

\item{extra_opts}{Extra options to feed to 3dresample (see afni docs for details).}
}
\value{
Resampled image (also saves as nii).
}
\description{
Resample image to a particular grid resolution.
}
