% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{rgb_to_grayscale}
\alias{rgb_to_grayscale}
\title{Convert RGB image to grayscale.}
\usage{
rgb_to_grayscale(image_path)
}
\arguments{
\item{image_path}{}
}
\value{
NA
}
\description{
Convert RGB image to grayscale.
}
