% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_model_params}
\alias{set_model_params}
\title{Set model parameters.}
\usage{
set_model_params(
  input_shape,
  num_classes,
  loss,
  activation,
  optimizer,
  metrics,
  ...
)
}
\arguments{
\item{input_shape}{4D (x,y,z,channel) vector for input dimensions.}

\item{num_classes}{Number of classes used for classification.}

\item{loss}{Loss function (in quotes, e.g. "binary_crossentropy")}

\item{activation}{Activation function (in quotes, e.g. "sigmoid")}

\item{optimizer}{Optimizer as a keras function (e.g. keras::optimizer_adam())}

\item{metrics}{Evaluation metrics; if more than 1, then make list.}

\item{...}{Any additional parameters for specific models.}
}
\value{
List of model parameters.
}
\description{
Set model parameters.
}
