% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{train_model_kfold}
\alias{train_model_kfold}
\title{Training with K-fold cross-validation.}
\usage{
train_model_kfold(
  model_building_fn,
  model_params,
  n_epochs,
  batch_size,
  dataset,
  n_folds
)
}
\arguments{
\item{n_folds}{}
}
\value{

}
\description{
Training with K-fold cross-validation.
}
