% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_functions.R
\name{train_model_standard}
\alias{train_model_standard}
\title{Standard training function.}
\usage{
train_model_standard(
  model,
  x_train,
  y_train,
  model_weights = NULL,
  epochs,
  batch_size,
  validation_data = NULL,
  validation_split = NULL,
  callbacks = NULL,
  class_weight = NULL,
  steps_per_epoch = NULL,
  validation_steps = NULL
)
}
\arguments{
\item{validation_steps}{}
}
\value{

}
\description{
Standard training function.
}
