% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni_functions.R
\name{extract_highest_diffusion_shell}
\alias{extract_highest_diffusion_shell}
\title{Extract highest shell of diffusion dataset.}
\usage{
extract_highest_diffusion_shell(in_file, out_file)

extract_highest_diffusion_shell(in_file, out_file)

extract_highest_diffusion_shell(in_file, out_file)
}
\arguments{
\item{out_file}{}
}
\value{
Nii object of highest shell image.

Nii object of highest shell image.

Nii object of highest shell image.
}
\description{
Extract highest shell of diffusion dataset.

Extract highest shell of diffusion dataset.

Extract highest shell of diffusion dataset.
}
