% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{resample_image}
\alias{resample_image}
\title{Resampling fn.}
\usage{
resample_image(data, target_dims = NULL, scale_factor = NULL)
}
\arguments{
\item{scale_factor}{}
}
\value{

}
\description{
Resampling fn.
}
